requirements.txt
# Django & REST Framework
Django>=5.0,<5.2
djangorestframework>=3.15.0
django-filter>=24.0
django-cors-headers>=4.4.0

# Database
psycopg2-binary>=2.9.9  # Django's PostgreSQL adapter
psycopg>=3.2.0  # Modern PostgreSQL adapter for LangGraph checkpointer
psycopg-pool>=3.2.0  # Connection pooling for psycopg
dj-database-url>=2.2.0
django-db-connection-pool>=1.2.5

# Configuration
python-decouple>=3.8
python-dotenv>=1.0.1

# Server
gunicorn>=22.0.0
daphne>=4.1.0  # ASGI server for Django Channels

# Authentication
django-allauth>=64.0.0
dj-rest-auth>=6.0.0
djangorestframework-simplejwt>=5.3.1
cryptography>=43.0.0

# AI/ML - Core Providers
openai>=1.50.0
anthropic>=0.40.0
google-generativeai==0.8.5
google-genai==1.52.0  # New Google AI SDK for image generation

# AI/ML - LangChain Ecosystem
langchain==0.3.27
langchain-openai==0.3.35
langchain-anthropic==0.3.22
langchain-google-genai==2.0.10
langgraph>=0.2.0
langgraph-checkpoint-redis>=0.2.0
langgraph-checkpoint-postgres>=3.0.1
langsmith>=0.1.140

# Vector Database
weaviate-client>=3.26.0,<4.0.0  # Vector database for personalization (v3 API)

# Task Queue
celery>=5.4.0
redis>=5.2.0

# WebSocket Support
channels>=4.1.0  # Django Channels for WebSocket
channels-redis>=4.2.1  # Redis backend for Channels
websockets>=13.0  # WebSocket client for testing

# Payments
stripe>=11.0.0

# Storage
minio>=7.2.9
boto3>=1.35.0

# Security
bleach>=6.1.0
django-ratelimit>=4.1.0
django-csp>=3.8
defusedxml>=0.7.1
better-profanity==0.7.0

# Media/Content
Pillow>=10.4.0
python-magic>=0.4.27
markdown>=3.7

# HTTP/API
requests>=2.32.0
httpx>=0.27.0
tenacity>=9.0.0  # Retry logic for API calls

# MCP & YAML
fastmcp>=2.0.0  # MCP client for multi-source analysis
pyyaml>=6.0.2  # For parsing docker-compose.yml files

# Monitoring
django-prometheus>=2.4.1

# Code Quality (dev)
ruff>=0.7.0
bandit[toml]>=1.7.10

# Testing (dev)
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0  # Async test support for LangGraph agent tests
